backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.3
rq==1.15.1
httpx==0.27.0
Pillow==10.3.0
